==> Capstone-Airbnb-NY/code/01_capstone.ipynb <==
# Capstone - Airbnb in New York

## Problem Statement

This project uses an Airbnb dataset to improve revenue management of Airbnb in NYC. The project utilizes the listings and reviews datasets to make inferences, practice NLP, and make predictions on the demand for accommodation in the Airbnb's and the best price-value ratio for guests.

## Contents:
- [Background](#Background)
- [Data Collection](./02_data_collection.ipynb)
- [Data Cleaning](./03_data_cleaning.ipynb)
- [Exploratory Data Analysis](./04_eda.ipynb)
- [Modeling](./05_modeling.ipynb)
- [Conclusions and Recommendations](./06_conclusions_and_recs.ipynb)

## Background

### Datasets

* [`listings.csv`](./data/listings.csv): Listing attributes for New York Airbnbs
* [`reviews.csv`](./data/reviews.csv): Reviews left by clients who stayed at the New York listings

### Data Dictionary

|Feature|Type|Dataset|Description|
|---|---|---|---|
|listing_id|float|reviews|The unique ID for each listing|
|id|object|float|The ID number|
|date|object|reviews|The data each review was written|
|reviewer_id|float|reviews|The unique ID for each reviewer|
|reviewer_name|object|reviews|The name of each reviewer|
|comments|object|reviews|The reviews left by customers who stayed in each listing|
|reviews_per_month|float|listings|The number of reviews per each month|
|number_of_reviews|float|listings|The total number of reviews|
|number_of_reviews_ltm|float|listings|The total number of reviews in the last twelve months|
|number_of_reviews_l30d|float|listings|The total number of reviews in the last 30 days|
|bedrooms|float|listings|The total number of bedrooms in each listing|
|accommodates|float|listings|The total number of people each listing can accommodate|
