==> Capstone-Airbnb-NY/code/02_data_collection.ipynb <==
# Data Collection

### Imports

import pandas as pd

### Data Import

# read in listings + reviews datasets
listings_df = pd.read_csv('../data/listings.csv')
reviews_df = pd.read_csv('../data/reviews.csv')

# check
listings_df.head(1)

reviews_df.head(1)

listings_df.columns

reviews_df.columns

listings_df.to_csv('../data/listings.csv')

reviews_df.to_csv('../data/reviews.csv')
