==> Capstone-Airbnb-NY/code/03_data_cleaning.ipynb <==
# Data Cleaning

### Imports

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction import text
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

import re

### Cleaning

# read in saved listings + reviews datasets from data collection notebook
listings_df = pd.read_csv('../data/listings.csv').drop(columns = 'Unnamed: 0')
reviews_df = pd.read_csv('../data/reviews.csv').drop(columns = 'Unnamed: 0')

#### Listings DF

# check first
listings_df.head(1)

# check last
listings_df.tail(1)

listings_df.dtypes

listings_df.shape

# check for missing values
listings_df.isnull().sum()

##### Note: A fair amount of missing values in reviews_per_month. The other missing values are insignificant to the entirety of the dataset. Can fill missing values in with 0's and ' '.

listings_df['host_acceptance_rate'].isnull().sum()

listings_df.info()

##### Check what missing values need to be 0's and which ones should be ' '

float_df = listings_df.select_dtypes(include = float, exclude = object)

listings_df.select_dtypes(include = float, exclude = object).columns

object_df = listings_df.select_dtypes(include = object, exclude = float)

listings_df.select_dtypes(include = object, exclude = float).columns

# replace missing object values with ' '
object_df.fillna(' ', inplace = True)

# check
object_df.isnull().sum()

object_df.shape

# replace missing float values with 0
float_df.fillna(0, inplace = True)

# check
float_df.isnull().sum()

float_df.shape

##### Now concatenate object_df and float_df back into the listings_df since there are no more missing values.

listings_df = pd.concat([object_df,float_df], axis = 1) # left/right

# check
listings_df.isnull().sum()

##### Save to cleaned csv

# save
listings_df.to_csv('../data/cleaned_listings.csv')

--- 

#### Reviews DF

# check first 5
reviews_df.head()

# check last 5
reviews_df.tail()

reviews_df.dtypes

reviews_df.shape

# check for missing values
reviews_df.isnull().sum()

##### Do the same thing as listings - separate into float and objects to replace with either ' ' or 0

float_reviews_df = reviews_df.select_dtypes(include = float, exclude = object)

float_reviews_df.select_dtypes(include = float, exclude = object).columns

object_reviews_df = reviews_df.select_dtypes(include = object, exclude = float)

object_reviews_df.select_dtypes(include = object, exclude = float).columns

# replace missing object values with ' '
object_reviews_df.fillna(' ', inplace = True)

# check
object_reviews_df.isnull().sum()

object_reviews_df.shape

# replace missing float values with 0
float_reviews_df.fillna(0, inplace = True)

# check
float_reviews_df.isnull().sum()

float_reviews_df.shape

##### Concatenate back together as reviews_df:

reviews_df = pd.concat([object_reviews_df,float_reviews_df], axis = 1) # left/right

# check
reviews_df.isnull().sum()

##### Save to cleaned csv

# save
reviews_df.to_csv('../data/cleaned_reviews.csv')
