==> Capstone-Airbnb-NY/code/05_modeling.ipynb <==
# Modeling

## Imports

import pandas as pd
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.pipeline import Pipeline
from sklearn.naive_bayes import MultinomialNB
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import confusion_matrix, ConfusionMatrixDisplay, plot_confusion_matrix

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer

from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

from sklearn.neighbors import KNeighborsClassifier
from sklearn.preprocessing import StandardScaler

from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import RandomForestClassifier, StackingClassifier

from sklearn.compose import ColumnTransformer

# read in cleaned data again
listings_df = pd.read_csv('../data/cleaned_listings.csv').drop(columns = 'Unnamed: 0')

listings_df.head(1)

listings_df['price'].dtypes

listings_df.shape

listings_df['price'].value_counts()

listings_df['price'] = listings_df['price'].replace({'\$':''}, regex = True)

listings_df['price'].head()

listings_df['price'] = listings_df['price'].replace({',':''}, regex = True)

listings_df['price'].head()

listings_df = listings_df[listings_df['price'] != ' ']

listings_df.iloc[[7551]]

listings_df = listings_df.drop(index=listings_df.iloc[7551].name)

listings_df = listings_df[listings_df['price'] != '2022-12-05']

listings_df['price'] = pd.to_numeric(listings_df['price'])

float_columns = listings_df.select_dtypes(float).columns

float_columns = list(float_columns)

float_columns.remove('price')

# set up X and y
X = listings_df[float_columns]
y = listings_df['price']

# train/test split
X_train, X_test, y_train, y_test = train_test_split(X,
                                                    y,
                                                    random_state=42)

lr = LinearRegression()

lr.fit(X_train, y_train)

lr.score(X_train, y_train)

lr.score(X_test, y_test)

# Lasso
sc = StandardScaler()

X_scaled = sc.fit_transform(X_train)

X_test_scaled = sc.transform(X_test)

lasso = Lasso()

lasso.fit(X_scaled, y_train)

lasso.score(X_scaled, y_train)

lasso.score(X_test_scaled, y_test)

# Ridge
sc = StandardScaler()

X_scaled = sc.fit_transform(X_train)

X_test_scaled = sc.transform(X_test)

ridge = Ridge()

ridge.fit(X_scaled, y_train)

ridge.score(X_scaled, y_train)

ridge.score(X_test_scaled, y_test)
