==> Capstone-Airbnb-NY/code/06_conclusions_and_recs.ipynb <==
# Conclusions and Recommendations

##### My models did not perform very well, indicating that predicting price of Airbnb listings in NY based off of other numeric features is not that helpful.

##### The highest score I got was with Linear Regression and hit only about 22% on my test set.

##### In terms of revenue management and predicting the demand for accommodations, it's important to consider New York's size. This means that in order to be within walking distance of popular attractions or at a reasonable price, the listing will ideally accomodate 2-4 people with 1-2 bedrooms.

##### Still working with the Listings dataset, but focusing on number of reviews, we can see that the total number of reviews and the number of reviews in the last twelve months is relatively similar, both gravitating towards 100-125. This is pretty on track with the number of reviews in the last 30 days with the average being around 10 reviews.

##### These ask the questions:
* which months are most likely to generate reviews? 
* Is there correlation for volume of bookings per month and number of reviews left per month?

##### But on the NLP side and switching to the Reviews dataset, I found the bigrams most commonly listed in the comments on the listings:
* new york
* great location
* highly recommend
* place stay
* great host
* great place
* definitely stay
* walking distance
* central park

##### These bigrams would beneficial to consider when promoting the listings, i.e. How far away is your listing from Central Park? Is it within walking distance of popular New York attractions? And lastly, it's important to consider how to display for your most positive, and popular reviews to potential or returner customers to show them you have superhost qualities.
